% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{as_tulpa_mesh}
\alias{as_tulpa_mesh}
\alias{as_tulpa_mesh.fm_mesh_2d}
\alias{as_tulpa_mesh.inla.mesh}
\title{Convert to a tulpa_mesh Object}
\usage{
as_tulpa_mesh(x, ...)

\method{as_tulpa_mesh}{fm_mesh_2d}(x, ...)

\method{as_tulpa_mesh}{inla.mesh}(x, ...)
}
\arguments{
\item{x}{Object to convert.}

\item{...}{Additional arguments (currently unused).}
}
\value{
A \code{tulpa_mesh} object.
}
\description{
Generic function to convert mesh objects from other packages into
\code{tulpa_mesh} objects. Currently supports \code{fm_mesh_2d} objects from
the fmesher package.
}
